requirements.txt
requests
beautifulsoup4
python-mpv
colorama
